Add UART adapter PCB
This PCB routes the UART0 pins on the 3π+ 2040's 2x7 Mid Expansion
Header out to a 6-pin FTDI compatible header. This allows for a FTDI
adapter to be used on the robot to check the UART signals from the GP28
and GP29 pins. I also plan to use this header configuration for getting
power and serial connectivity to my MRI-SWD debug probe in the future.
Having access to the serial pins will allow it to expose the serial
traffic on another WiFi socket.

I updated the README as well to give an overview of the progress I have
made on MRI-SWD and this adapter PCB.